fix: changed telegramId type

+++ MovieRecommendationBot/src/main/java/ru/spbstu/movierecbot/dao/GenreDao.java
@@ -4,11 +4,11 @@ import java.util.List;
 
 public interface GenreDao {
     //Добавление жанра предпочтения пользователя
-    int addGenre(int telegramId, String name);
+    int addGenre(long telegramId, String name);
 
     //Получение всех жанров из предпочтений пользователя
-    List<GenresRecord> getGenresByTelegramId(int telegramId);
+    List<GenresRecord> getGenresByTelegramId(long telegramId);
 
     //Удаление жанра из предпочтений
-    int deleteGenre(int telegramId, String genre);
+    int deleteGenre(long telegramId, String genre);
 }
